+++ src/routes/ConsortiumManager/settings/data-import/DataImportLogs/index.js
@@ -1,0 +1,1 @@
+export { DataImportLogs } from './DataImportLogs';