+++ src/routes/ConsortiumManager/settings/data-import/DataImportSettings.js
@@ -1,20 +1,9 @@
-import { FormattedMessage } from 'react-intl';
+import { DataImportLogs } from './DataImportLogs';
 
-import { stripesShape } from '@folio/stripes/core';
-import { Settings } from '@folio/stripes/smart-components';
-
-const DataImportSettings = (props) => {
+const DataImportSettings = () => {
   return (
-    <Settings
-      {...props}
-      sections={[]}
-      paneTitle={<FormattedMessage id="ui-data-import.settings.index.paneTitle" />}
-    />
+    <DataImportLogs />
   );
-};
-
-DataImportSettings.propTypes = {
-  stripes: stripesShape.isRequired,
 };
 
 export default DataImportSettings;