+++ CHANGELOG.md
@@ -13,3 +13,4 @@
 * [UICONSET-8](https://issues.folio.org/browse/UICONSET-8) Limit access to "Consortium manager" interface.
 * [UICONSET-10](https://issues.folio.org/browse/UICONSET-10) Select member institutions for consortia management.
 * [UICONSET-26](https://issues.folio.org/browse/UICONSET-26) Set consortium manager settings section.
+* [UICONSET-123](https://issues.folio.org/browse/UICONSET-123) View list of import jobs.